+++ rich/json.py
@@ -1,4 +1,5 @@
 from json import loads, dumps
+from typing import Any
 
 from .text import Text
 from .highlighter import JSONHighlighter, NullHighlighter
@@ -13,13 +14,31 @@ class JSON:
         highlight (bool, optional): Enable highlighting. Defaults to True.
     """
 
-    def __init__(self, json: str, indent: int = 4, highlight: bool = True) -> None:
+    def __init__(self, json: str, indent: int = 2, highlight: bool = True) -> None:
         data = loads(json)
         json = dumps(data, indent=indent)
         highlighter = JSONHighlighter() if highlight else NullHighlighter()
         self.text = highlighter(json)
         self.text.no_wrap = True
         self.text.overflow = None
+
+    @classmethod
+    def from_data(cls, data: Any, indent: int = 2, highlight: bool = True) -> "JSON":
+        """Encodes a JSON object from arbitrary data.
+
+        Returns:
+            Args:
+                data (Any): An object that may be encoded in to JSON
+                indent (int, optional): Number of characters to indent by. Defaults to True.
+                highlight (bool, optional): Enable highlighting. Defaults to True.
+        """
+        json_instance: "JSON" = cls.__new__(cls)
+        json = dumps(data, indent=indent)
+        highlighter = JSONHighlighter() if highlight else NullHighlighter()
+        json_instance.text = highlighter(json)
+        json_instance.text.no_wrap = True
+        json_instance.text.overflow = None
+        return json_instance
 
     def __rich__(self) -> Text:
         return self.text
@@ -36,6 +55,14 @@ if __name__ == "__main__":
         metavar="PATH",
         help="path to file, or - for stdin",
     )
+    parser.add_argument(
+        "-i",
+        "--indent",
+        metavar="SPACES",
+        type=int,
+        help="Number of spaces in an indent",
+        default=2,
+    )
     args = parser.parse_args()
 
     from rich.console import Console
@@ -50,4 +77,4 @@ if __name__ == "__main__":
         error_console.print(f"Unable to read {args.path!r}; {error}")
         sys.exit(-1)
 
-    console.print(JSON(json_data), soft_wrap=True)
+    console.print(JSON(json_data, indent=args.indent), soft_wrap=True)