+++ tests/test_console.py
@@ -119,7 +119,22 @@ def test_print():
 
 def test_print_json():
     console = Console(file=io.StringIO(), color_system="truecolor")
-    console.print_json('[false, true, null, "foo"]')
+    console.print_json('[false, true, null, "foo"]', indent=4)
+    result = console.file.getvalue()
+    print(repr(result))
+    expected = '\x1b[1m[\x1b[0m\n    \x1b[3;91mfalse\x1b[0m,\n    \x1b[3;92mtrue\x1b[0m,\n    \x1b[3;35mnull\x1b[0m,\n    \x1b[32m"foo"\x1b[0m\n\x1b[1m]\x1b[0m\n'
+    assert result == expected
+
+
+def test_print_json_error():
+    console = Console(file=io.StringIO(), color_system="truecolor")
+    with pytest.raises(TypeError):
+        console.print_json(["foo"], indent=4)
+
+
+def test_print_json_data():
+    console = Console(file=io.StringIO(), color_system="truecolor")
+    console.print_json(data=[False, True, None, "foo"], indent=4)
     result = console.file.getvalue()
     print(repr(result))
     expected = '\x1b[1m[\x1b[0m\n    \x1b[3;91mfalse\x1b[0m,\n    \x1b[3;92mtrue\x1b[0m,\n    \x1b[3;35mnull\x1b[0m,\n    \x1b[32m"foo"\x1b[0m\n\x1b[1m]\x1b[0m\n'