+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.9.0] - 2020-08-29
+
+### Added
+
+- Added data parameter to print_json method / function
+
+### Changed
+
+- Changed default indent of JSON to 2 (down from 4)
+- Changed highlighting of JSON keys to new style (bold blue)
+
 ## [10.8.0] - 2020-08-28
 
 ### Added